backend/requirements.txt
fastapi
uvicorn
imapclient
pyzmail36
smtplib
sqlalchemy
aiosqlite